# server/qualiva-openapi.yaml
openapi: 3.0.3
info:
  title: Qualiva - Enterprise Secure Module Management System (ESMMS) API
  description: |
    Comprehensive OpenAPI 3.x specification for the Qualiva ESMMS backend services.
    Generated strictly via comprehensive source code inspection of `server/src` for API security auditing with OWASP ZAP.
    
    Includes role-based access control (RBAC), multi-engine database proxies, granular moderator permissions,
    proctored assessment workflows, and real-time enterprise communication endpoints.
  version: 1.0.0
servers:
  - url: http://host.docker.internal:5050
    description: Local development backend server
  - url: /
    description: Current host relative path
...
paths:
...
  /api/v1/modules:
    get:
      tags:
        - Course Modules
      summary: List Course Modules (Public Catalog)
      description: Returns active curriculum modules with caching support.
      operationId: listModules
      parameters:
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, DRAFT, ARCHIVED, DEPRECATED]
            default: ACTIVE
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [ACADEMY_COURSE, WORK_MODULE]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Module catalog list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  modules:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModuleSummary'
    post:
      tags:
        - Course Modules
      summary: Create Course Module
      description: Provision a new curriculum course module in `DRAFT` status.
      operationId: createModule
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - title
                - department
              properties:
                code:
                  type: string
                  example: "CS-401"
                title:
                  type: string
                  example: "Advanced Cloud Security Engineering"
                description:
                  type: string
                department:
                  type: string
                  example: "Information Security"
                level:
                  type: string
                  enum: [Beginner, Intermediate, Advanced]
                  default: Beginner
                duration:
                  type: string
                  example: "8 Weeks"
                type:
                  type: string
                  enum: [ACADEMY_COURSE, WORK_MODULE]
                  default: ACADEMY_COURSE
                outcomes:
                  type: array
                  items:
                    type: string
                prerequisites:
                  type: array
                  items:
                    type: string
                expiresMonths:
                  type: integer
                  default: 12
      responses:
        '201':
          description: Module created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  module:
                    $ref: '#/components/schemas/ModuleDetail'
        '400':
          description: Missing code, title, or department.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Module code already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/modules/{id}:
    get:
      tags:
        - Course Modules
      summary: Get Single Module with Curriculum
      description: Returns full course metadata, sections, lessons, and assessment thresholds.
      operationId: getModuleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full curriculum retrieved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  module:
                    $ref: '#/components/schemas/ModuleDetail'
        '404':
          description: Module not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Course Modules
      summary: Update Course Module
      description: Updates course title, description, outcomes, prerequisites, or status.
      operationId: updateModule
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Module updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  module:
                    $ref: '#/components/schemas/ModuleDetail'
        '403':
          description: Forbidden role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Course Modules
      summary: Delete Course Module
      description: Deletes a module and cascades to all child sections and lessons.
      operationId: deleteModule
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '403':
          description: Forbidden role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/assessments:
    get:
      tags:
        - Assessments & Exams
      summary: List Question Banks & Assessments
      description: Creator and Admin overview of assessment question banks.
      operationId: listAssessments
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      parameters:
        - name: moduleId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Assessment banks retrieved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  assessments:
                    type: array
                    items:
                      type: object
    post:
      tags:
        - Assessments & Exams
      summary: Create Assessment Question Bank
      description: Configures new assessment with questions, passing threshold, and time limit.
      operationId: createAssessment
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - moduleId
                - title
              properties:
                moduleId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                passingScore:
                  type: integer
                  default: 70
                sampleSize:
                  type: integer
                  default: 10
                durationMinutes:
                  type: integer
                  default: 45
                randomizeQuestions:
                  type: boolean
                  default: true
                questions:
                  type: array
                  items:
                    $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Assessment created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  assessment:
                    type: object
  /api/v1/assessments/{id}:
    get:
      tags:
        - Assessments & Exams
      summary: Get Assessment Details with Questions
      description: Returns full assessment definition including correct answers for creator editing.
      operationId: getAssessmentById
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Assessment definition retrieved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  assessment:
                    type: object
        '404':
          description: Assessment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Assessments & Exams
      summary: Update Assessment Question Bank
      description: Update thresholds, time limit, question pools, or status.
      operationId: updateAssessment
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Assessment updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  assessment:
                    type: object
    delete:
      tags:
        - Assessments & Exams
      summary: Delete Assessment Question Bank
      description: Permanently removes assessment and associated submission records.
      operationId: deleteAssessment
      security:
        - bearerAuth: []
      x-required-roles:
        - COURSE_CREATOR
        - ADMIN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Assessment deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'

  # =========================================================================
  # Cryptographic Certificates
  # =========================================================================
...
  /api/v1/admin/users:
    get:
      tags:
        - Admin Management
      summary: List Enterprise User Directory
      description: Paginated directory of platform accounts with status, department, and role filtering.
      operationId: adminListUsers
      security:
        - bearerAuth: []
      x-required-roles:
        - ADMIN
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: User directory page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDetail'
                  pagination:
                    type: object
                  stats:
                    type: object
    post:
      tags:
        - Admin Management
      summary: Provision Enterprise User Account
      description: Manually creates an employee account with a temporary password and forces reset upon first login.
      operationId: adminCreateUser
      security:
        - bearerAuth: []
      x-required-roles:
        - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                role:
                  type: string
                  enum: [USER, COURSE_CREATOR, MODERATOR, ADMIN]
                  default: USER
                department:
                  type: string
                temporaryPassword:
                  type: string
                phone:
                  type: string
                location:
                  type: string
      responses:
        '201':
          description: User provisioned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/UserSummary'
        '409':
          description: Email already in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/moderator/users:
    get:
      tags:
        - Moderator Operations
      summary: List Users (Moderator Portal)
      description: Read-only user directory for moderators. Administrator accounts are strictly excluded from output.
      operationId: moderatorListUsers
      security:
        - bearerAuth: []
      x-required-moderator-permission:
        module: users
        action: view
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Users list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDetail'
        '403':
          description: Permission denied (lacking users.view).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/moderator/enrollments:
    get:
      tags:
        - Moderator Operations
      summary: View Enrollment Approvals Queue
      description: Moderator access to view pending enrollment applications forwarded by creators.
      operationId: moderatorGetEnrollments
      security:
        - bearerAuth: []
      x-required-moderator-permission:
        module: enrollments
        action: view
      responses:
        '200':
          description: Enrollment queue.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/EnrollmentRequest'
        '403':
          description: Permission denied (lacking enrollments.view).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/chat/notifications:
    get:
      tags:
        - Real-Time Chat & Inbox
      summary: Get Mailbox Dropdown Feed
      description: Returns combined preview list of recent announcements, support tickets, and direct messages for the top navigation bar.
      operationId: getChatNotificationsFeed
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification feed items.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  announcements:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
                  community:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
                  direct:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'

  # =========================================================================
  # System Notifications
  # =========================================================================
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: List Persistent Notifications
      description: Returns in-app notifications targeted to the caller's user ID or RBAC role.
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Notifications list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  unreadCount:
                    type: integer
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
...
  /api/v1/transfers:
    get:
      tags:
        - Transfers & SLA
      summary: List Course Transfer & SLA Requests
      description: Lists transfer and SLA extension requests with status counts.
      operationId: listTransfers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, PENDING, APPROVED, REJECTED]
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, COURSE_TRANSFER, SLA_EXTENSION, DEPARTMENT_TRANSFER]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Transfers list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  counts:
                    type: object
                    properties:
                      pending:
                        type: integer
                      approved:
                        type: integer
                      rejected:
                        type: integer
                      total:
                        type: integer
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransferRequest'
    post:
      tags:
        - Transfers & SLA
      summary: Submit Course Transfer or SLA Extension Request
      description: Student submits request to change tracks, request SLA deadline extension, or relocate department.
      operationId: createTransfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                type:
                  type: string
                  enum: [COURSE_TRANSFER, SLA_EXTENSION, DEPARTMENT_TRANSFER]
                  default: COURSE_TRANSFER
                courseId:
                  type: string
                targetCourseId:
                  type: string
                fromDepartment:
                  type: string
                toDepartment:
                  type: string
                reason:
                  type: string
                  example: "Requesting 14-day SLA deadline extension due to medical leave."
                requestedDays:
                  type: integer
                  default: 14
      responses:
        '201':
          description: Transfer request created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  transfer:
                    $ref: '#/components/schemas/TransferRequest'
        '400':
          description: Reason is mandatory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JSON Web Token passed in the `Authorization` header as `Bearer <token>`.
        Payload contains `sub` (User UUID), `email`, `role`, `status`, and `department`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: esmms_refresh_token
      description: |
        HttpOnly cookie containing the cryptographic refresh token used exclusively by `/api/v1/auth/refresh` and `/api/v1/auth/logout`.
  schemas:
    StandardSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully."
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request payload."
        details:
          type: array
          items:
            type: string
...
    ModuleSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: "SEC-101"
        title:
          type: string
        description:
          type: string
          nullable: true
        department:
          type: string
        level:
          type: string
        duration:
          type: string
          nullable: true
        instructorName:
          type: string
          nullable: true
        type:
          type: string
        status:
          type: string
        rating:
          type: number
        ratingCount:
          type: integer
        outcomes:
          type: array
          items:
            type: string
        prerequisites:
          type: array
          items:
            type: string
        expiresMonths:
          type: integer
        expiresAt:
          type: string
          format: date-time
          nullable: true
        sectionCount:
          type: integer
        enrolledCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ModuleDetail:
      allOf:
        - $ref: '#/components/schemas/ModuleSummary'
        - type: object
          properties:
            sections:
              type: array
              items:
                $ref: '#/components/schemas/Section'
    Section:
      type: object
      properties:
        id:
          type: string
          format: uuid
        moduleId:
          type: string
        title:
          type: string
        order:
          type: integer
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
...